Make secrets and key management public (#645)
* Move secrets out of internal/

* Add note about public

* Move signing public, rename to keys

* Finish rename

* Rename main file

* Cleanup package comments
