+++ internal/admin/config.go
@@ -23,10 +23,10 @@ import (
 	"net/http"
 
 	"github.com/google/exposure-notifications-server/internal/database"
-	"github.com/google/exposure-notifications-server/internal/secrets"
 	"github.com/google/exposure-notifications-server/internal/setup"
-	"github.com/google/exposure-notifications-server/internal/signing"
 	"github.com/google/exposure-notifications-server/internal/storage"
+	"github.com/google/exposure-notifications-server/pkg/keys"
+	"github.com/google/exposure-notifications-server/pkg/secrets"
 )
 
 var _ setup.BlobstoreConfigProvider = (*Config)(nil)
@@ -36,7 +36,7 @@ var _ setup.SecretManagerConfigProvider = (*Config)(nil)
 
 type Config struct {
 	Database      database.Config
-	KeyManager    signing.Config
+	KeyManager    keys.Config
 	SecretManager secrets.Config
 	Storage       storage.Config
 
@@ -50,7 +50,7 @@ func (c *Config) DatabaseConfig() *database.Config {
 	return &c.Database
 }
 
-func (c *Config) KeyManagerConfig() *signing.Config {
+func (c *Config) KeyManagerConfig() *keys.Config {
 	return &c.KeyManager
 }
