+++ internal/export/config.go
@@ -20,10 +20,10 @@ import (
 
 	"github.com/google/exposure-notifications-server/internal/database"
 	"github.com/google/exposure-notifications-server/internal/observability"
-	"github.com/google/exposure-notifications-server/internal/secrets"
 	"github.com/google/exposure-notifications-server/internal/setup"
-	"github.com/google/exposure-notifications-server/internal/signing"
 	"github.com/google/exposure-notifications-server/internal/storage"
+	"github.com/google/exposure-notifications-server/pkg/keys"
+	"github.com/google/exposure-notifications-server/pkg/secrets"
 )
 
 // Compile-time check to assert this config matches requirements.
@@ -37,7 +37,7 @@ var _ setup.ObservabilityExporterConfigProvider = (*Config)(nil)
 // the export components.
 type Config struct {
 	Database              database.Config
-	KeyManager            signing.Config
+	KeyManager            keys.Config
 	SecretManager         secrets.Config
 	Storage               storage.Config
 	ObservabilityExporter observability.Config
@@ -61,7 +61,7 @@ func (c *Config) DatabaseConfig() *database.Config {
 	return &c.Database
 }
 
-func (c *Config) KeyManagerConfig() *signing.Config {
+func (c *Config) KeyManagerConfig() *keys.Config {
 	return &c.KeyManager
 }
