+++ internal/serverenv/env.go
@@ -24,9 +24,9 @@ import (
 	"github.com/google/exposure-notifications-server/internal/database"
 	"github.com/google/exposure-notifications-server/internal/metrics"
 	"github.com/google/exposure-notifications-server/internal/observability"
-	"github.com/google/exposure-notifications-server/internal/secrets"
-	"github.com/google/exposure-notifications-server/internal/signing"
 	"github.com/google/exposure-notifications-server/internal/storage"
+	"github.com/google/exposure-notifications-server/pkg/keys"
+	"github.com/google/exposure-notifications-server/pkg/secrets"
 )
 
 // ExporterFunc defines a factory function for creating a context aware metrics exporter.
@@ -38,7 +38,7 @@ type ServerEnv struct {
 	blobstore             storage.Blobstore
 	database              *database.DB
 	exporter              metrics.ExporterFromContext
-	keyManager            signing.KeyManager
+	keyManager            keys.KeyManager
 	secretManager         secrets.SecretManager
 	observabilityExporter observability.Exporter
 }
@@ -95,7 +95,7 @@ func WithSecretManager(sm secrets.SecretManager) Option {
 }
 
 // WithKeyManager creates an Option to install a specific KeyManager to use for signing requests.
-func WithKeyManager(km signing.KeyManager) Option {
+func WithKeyManager(km keys.KeyManager) Option {
 	return func(s *ServerEnv) *ServerEnv {
 		s.keyManager = km
 		return s
@@ -122,7 +122,7 @@ func (s *ServerEnv) SecretManager() secrets.SecretManager {
 	return s.secretManager
 }
 
-func (s *ServerEnv) KeyManager() signing.KeyManager {
+func (s *ServerEnv) KeyManager() keys.KeyManager {
 	return s.keyManager
 }
