+++ pkg/keys/keys.go
@@ -12,8 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package signing defines the interface to and implementation of signing
-package signing
+// Package keys defines the interface to and implementation of key management
+// operations.
+//
+// Although exported, this package is non intended for general consumption. It
+// is a shared dependency between multiple exposure notifications projects. We
+// cannot guarantee that there won't be breaking changes in the future.
+package keys
 
 import (
 	"context"
