+++ pkg/keys/noop.go
@@ -12,40 +12,24 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package signing
+package keys
 
 import (
 	"context"
 	"crypto"
 	"fmt"
-
-	"github.com/aws/aws-sdk-go/aws/session"
-	"github.com/aws/aws-sdk-go/service/kms"
-	"github.com/lstoll/awskms"
 )
 
 // Compile-time check to verify implements interface.
-var _ KeyManager = (*AWSKMS)(nil)
+var _ KeyManager = (*Noop)(nil)
 
-// AWSKMS implements the signing.KeyManager interface and can be used to sign
-// export files using AWS KMS.
-type AWSKMS struct {
-	svc *kms.KMS
+// Noop is a key manager that does nothing and always returns an error.
+type Noop struct{}
+
+func NewNoop(ctx context.Context) (KeyManager, error) {
+	return &Noop{}, nil
 }
 
-func NewAWSKMS(ctx context.Context) (KeyManager, error) {
-	sess, err := session.NewSession()
-	if err != nil {
-		return nil, fmt.Errorf("failed to create session: %w", err)
-	}
-
-	svc := kms.New(sess)
-
-	return &AWSKMS{
-		svc: svc,
-	}, nil
-}
-
-func (s *AWSKMS) NewSigner(ctx context.Context, keyID string) (crypto.Signer, error) {
-	return awskms.NewSigner(ctx, s.svc, keyID)
+func (n *Noop) NewSigner(ctx context.Context, keyID string) (crypto.Signer, error) {
+	return nil, fmt.Errorf("noop cannot sign")
 }
